+++ CMakeLists.txt
@@ -27,9 +27,9 @@ set(SHARED_SOURCES
     src/Core/Math/Math.cpp
     src/World/Entity.cpp
     src/World/WorldManager.cpp
-    src/Core/Networking/FinalverseClient.cpp
-    src/Core/Networking/MessageProtocol.cpp
-    src/Core/Networking/NetworkClient.cpp
+    src/Network/FinalverseClient.cpp
+    src/Network/MessageProtocol.cpp
+    src/Network/NetworkClient.cpp
     src/Core/Audio/AudioEngine.cpp
     src/Core/Audio/SpatialAudioSystem.cpp
     src/Core/Input/InteractionManager.cpp
