+++ src/Network/FinalverseClient.h
@@ -14,7 +14,7 @@
 #include <queue>
 #include <mutex>
 #include "Core/Math/Math.h"
-#include "Core/Networking/MessageProtocol.h"
+#include "Network/MessageProtocol.h"
 #include "Visualization/DataVisualizer.h"
 #include <websocketpp/client.hpp>
 #include <websocketpp/config/asio_no_tls_client.hpp>
