+++ verify_structure.sh
@@ -12,10 +12,10 @@ files=(
     "src/World/Entity.cpp"
     "src/World/WorldManager.h"
     "src/World/WorldManager.cpp"
-    "src/Core/Networking/FinalverseClient.h"
-    "src/Core/Networking/FinalverseClient.cpp"
-    "src/Core/Networking/MessageProtocol.h"
-    "src/Core/Networking/MessageProtocol.cpp"
+    "src/Network/FinalverseClient.h"
+    "src/Network/FinalverseClient.cpp"
+    "src/Network/MessageProtocol.h"
+    "src/Network/MessageProtocol.cpp"
     "src/Core/Audio/AudioEngine.h"
     "src/Core/Audio/AudioEngine.cpp"
     "src/Rendering/Renderer.h"
